+++ lib/mastodon_api.dart
@@ -33,6 +33,8 @@ export 'package:mastodon_api/src/service/entities/instance_statuses_configuratio
 export 'package:mastodon_api/src/service/entities/instance_urls_configuration.dart';
 export 'package:mastodon_api/src/service/entities/instance_usage.dart';
 export 'package:mastodon_api/src/service/entities/instance_usage_users.dart';
+export 'package:mastodon_api/src/service/entities/poll.dart';
+export 'package:mastodon_api/src/service/entities/poll_option.dart';
 export 'package:mastodon_api/src/service/entities/rate_limit.dart';
 export 'package:mastodon_api/src/service/entities/registered_application.dart';
 export 'package:mastodon_api/src/service/entities/status.dart';